feat: Make instance profile optional; add name arn and role properties
The instance construct only takes a role not a full instance profile. So in most cases you likely dont need a full instance profile unless its created for another purpose. The default behavior of creating a instance profile remains the same, but is now optionally able to be disabled. In addition this adds other common properties to work with.

Closes #197